+++ libblobd-direct/src/exporter.rs
@@ -22,7 +22,7 @@ use std::sync::Arc;
 
 #[derive(PartialEq, Eq, Clone, Copy, Debug, Default, Serialize, Deserialize)]
 pub struct BlobdExporterMarker {
-  object_id: u64,
+  object_id: u128,
   partition: usize,
 }
 
@@ -44,7 +44,7 @@ impl PartialOrd for BlobdExporterMarker {
 
 pub struct BlobdExportedObject {
   // This is generally a blobd internal value and not that useful, but we'll export it anyway.
-  pub id: u64,
+  pub id: u128,
   pub key: Vec<u8>,
   pub size: u64,
   pub created: DateTime<Utc>,
